requirements.txt
streamlit
pandas
numpy
rembg
pillow
zipfile36